# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DentApp API",
        "description": "API aplikacji DentApp",
        "version": "1.0.0"
    },
    "paths": {
        "/api/announcement": {
            "get": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Pobiera listę ogłoszeń (jedno na dzień, dla wszystkich)",
                "operationId": "3c19ddbd61e53cfd67bf34a679aff1bc",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/AnnouncementResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Tworzy nowe ogłoszenie (jedno na dany dzień)",
                "operationId": "073f8bc8afaa696764e383a9fb659600",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "published_at"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Spotkanie zespołu"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Spotkanie odbędzie się o 10:00."
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utworzono",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AnnouncementResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Błąd walidacji (np. ogłoszenie na ten dzień już istnieje)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/announcement/{uuid}": {
            "get": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Pobiera szczegóły ogłoszenia",
                "operationId": "ef677031713a8295b7197aeae5f87029",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AnnouncementResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Aktualizuje ogłoszenie",
                "operationId": "7ed21e03321c727cd69f1577d4680f97",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "published_at"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AnnouncementResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Usuwa ogłoszenie",
                "operationId": "73a9586d1583cc65acb189dbf2e0dc7b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usunięto"
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/message": {
            "get": {
                "tags": [
                    "Message"
                ],
                "summary": "Skrzynka wiadomości zalogowanego użytkownika (odebrane, wysłane, broadcasty)",
                "operationId": "a7c6abc72ab4d49342f875a1643ecea4",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/MessageResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "Wysyła wiadomość do konkretnej osoby lub do grupy (wymagany dokładnie jeden cel)",
                "operationId": "ee15771b1dadd0efb04ebd263be4f168",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string"
                                    },
                                    "recipient_uuid": {
                                        "description": "Wiadomość bezpośrednia do użytkownika (wyklucza message_group_uuid)",
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "message_group_uuid": {
                                        "description": "Wiadomość do grupy (wyklucza recipient_uuid)",
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utworzono",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MessageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedResponse": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "from": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "private_email": {
                        "type": "string",
                        "format": "email"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "pwz_numer": {
                        "type": "string",
                        "nullable": true
                    },
                    "avatar_path": {
                        "type": "string",
                        "nullable": true
                    },
                    "background_path": {
                        "type": "string",
                        "nullable": true
                    },
                    "private_phone_number": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone_number": {
                        "type": "string",
                        "nullable": true
                    },
                    "job_positions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/JobPositionResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "JobPositionResource": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "MessageResource": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "recipient_user_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "message_group_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "message": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "AnnouncementResource": {
                "properties": {
                    "uuid": {
                        "description": "Summary of AnnouncementResource",
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "author": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        ],
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Sanctum — przekaż w nagłówku: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}